abstract out Premium and reorganize sections
