+++ web/molecules/sections/index.ts
@@ -1,0 +1,5 @@
+import { Faq } from './Faq';
+import { Features } from './Features';
+import { Premium } from './Premium';
+
+export { Faq, Features, Premium };
